+++ flutter/lib/src/web/local_media_track.dart
@@ -60,4 +60,65 @@ class WebLocalMediaTrack implements LocalMediaTrack {
     void fn(JSAny? l) => f((l as JSNumber).toDartInt);
     fallibleFunction(() => obj.on_audio_level_changed(fn.toJS));
   }
+
+  @override
+  bool isAudioProcessingAvailable() {
+    return fallibleFunction(() => obj.is_audio_processing_available());
+  }
+
+  @override
+  Future<void> setAutoGainControlEnabled(bool enabled) async {
+    await fallibleFuture(obj.set_auto_gain_control_enabled(enabled).toDart);
+  }
+
+  @override
+  Future<void> setEchoCancellationEnabled(bool enabled) async {
+    await fallibleFuture(obj.set_echo_cancellation_enabled(enabled).toDart);
+  }
+
+  @override
+  Future<void> setNoiseSuppressionEnabled(bool enabled) async {
+    await fallibleFuture(obj.set_noise_suppression_enabled(enabled).toDart);
+  }
+
+  @override
+  Future<bool> isAutoGainControlEnabled() async {
+    return (await fallibleFuture(
+      obj.is_auto_gain_control_enabled().toDart,
+    )).toDart;
+  }
+
+  @override
+  Future<bool> isEchoCancellationEnabled() async {
+    return (await fallibleFuture(
+      obj.is_echo_cancellation_enabled().toDart,
+    )).toDart;
+  }
+
+  @override
+  Future<bool> isNoiseSuppressionEnabled() async {
+    return (await fallibleFuture(
+      obj.is_noise_suppression_enabled().toDart,
+    )).toDart;
+  }
+
+  @override
+  Future<void> setNoiseSuppressionLevel(NoiseSuppressionLevel enabled) async {
+    throw UnsupportedError('setNoiseSuppressionLevel in not supported on web');
+  }
+
+  @override
+  Future<NoiseSuppressionLevel> getNoiseSuppressionLevel() async {
+    throw UnsupportedError('getNoiseSuppressionLevel in not supported on web');
+  }
+
+  @override
+  Future<void> setHighPassFilterEnabled(bool enabled) async {
+    throw UnsupportedError('setHighPassFilterEnabled in not supported on web');
+  }
+
+  @override
+  Future<bool> isHighPassFilterEnabled() async {
+    throw UnsupportedError('isHighPassFilterEnabled in not supported on web');
+  }
 }